+++ documentation/docs/changelog.md
@@ -2,6 +2,8 @@
 
 ## yyyy.mm.dd - version 1.4.0
 
+* Add the id property to the Item model
+
 ## 2024.05.24 - version 1.3.0
 
 * Documentation published on [GitHub Pages](https://www.stormwindlibrary.com)