+++ documentation/docs/changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## yyyy.mm.dd - version 1.4.0
+## 2024.06.11 - version 1.4.0
 
 * Add a factory to create item instances from multiple sources
 * Add new helper methods to Arr